requirements.txt
django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1
django-cors-headers